+++ Client/src/main/java/org/example/client/controller/filter/FilterListener.java
@@ -5,5 +5,5 @@ import java.time.Instant;
 public interface FilterListener {
     void onFiltersCleared();
 
-    void onFiltersApplied(String type, String category, Instant dateFrom, Instant dateTo, Integer minAmount, Integer maxAmount);
+    void onFiltersApplied(String type, String category, Instant dateFrom, Instant dateTo, Double minAmount, Double maxAmount);
 }
