+++ Common/src/main/java/org/example/dto/request/DecreaseOperationRequest.java
@@ -10,7 +10,7 @@ import java.time.Instant;
 public class DecreaseOperationRequest extends OperationRequest {
     private final DecreaseOperationCategory category;
 
-    public DecreaseOperationRequest(String token, int amount, Instant dateTime, DecreaseOperationCategory category) {
+    public DecreaseOperationRequest(String token, double amount, Instant dateTime, DecreaseOperationCategory category) {
         super(RequestAction.DECREASE_OPERATION, token, amount, dateTime);
         this.category = category;
     }
